backend/requirements.txt
fastapi==0.123.7
psycopg==3.3.1
pydantic-settings==2.12.0
SQLAlchemy==2.0.44
uvicorn==0.38.0
psycopg-binary==3.3.1
